infra/infra/scripts/data_scripts/requirements_load_tables_worked_locally.txt
# sucessfully installed for Python 3.11.9 on 2025-03-17
azure-common           1.1.28
azure-core             1.32.0
azure-identity         1.19.0
azure-keyvault-secrets 4.7.0
certifi                2025.1.31
cffi                   1.17.1
charset-normalizer     3.4.1
cryptography           44.0.2
idna                   3.10
isodate                0.7.2
msal                   1.32.0
msal-extensions        1.3.1
numpy                  2.2.4
pandas                 2.2.2
pip                    24.0
psycopg2-binary        2.9.10
pycparser              2.22
PyJWT                  2.10.1
python-dateutil        2.9.0.post0
pytz                   2025.1
requests               2.32.3
setuptools             65.5.0
six                    1.17.0
typing_extensions      4.12.2
tzdata                 2025.1
urllib3                2.3.0
